Move tests from test_tree_additional.py to test_tree.py
Consolidate all tree explainer tests into the standard test file to follow
repository conventions. All 28 additional tests for TreeExplainer coverage
are now in tests/explainers/test_tree.py.
